refactor: 서치바 CSS 수정 및 컴포넌트 이름 변경
